requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
diff-match-patch==20230430
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==0.63.6
django-import-export==4.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
mysqlclient==2.2.4
openpyxl==3.1.5
pillow==10.4.0
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
razorpay==1.4.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stripe==10.8.0
tablib==3.5.0
twilio==9.2.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
yarl==1.9.4
